rename sui module to iota
