+++ identity_iota_core/src/rebased/assets/asset.rs
@@ -5,7 +5,7 @@ use std::str::FromStr as _;
 
 use crate::rebased::client::IdentityClient;
 use crate::rebased::client::IotaKeySignature;
-use crate::rebased::sui::move_calls;
+use crate::rebased::iota::move_calls;
 use crate::rebased::transaction::Transaction;
 use crate::rebased::transaction::TransactionOutput;
 use crate::rebased::utils::MoveType;
