+++ identity_iota_core/src/rebased/iota/move_calls/asset/delete.rs
