+++ identity_iota_core/src/rebased/iota/move_calls/identity/borrow_asset.rs
@@ -15,7 +15,7 @@ use iota_sdk::types::transaction::ProgrammableTransaction;
 use move_core_types::ident_str;
 
 use crate::rebased::proposals::BorrowAction;
-use crate::rebased::sui::move_calls::utils;
+use crate::rebased::iota::move_calls::utils;
 use crate::rebased::utils::MoveType;
 
 pub(crate) fn propose_borrow(
