+++ identity_iota_core/src/rebased/iota/move_calls/identity/create.rs
@@ -12,7 +12,7 @@ use iota_sdk::types::IOTA_FRAMEWORK_PACKAGE_ID;
 use move_core_types::ident_str;
 
 use crate::rebased::migration::OnChainIdentity;
-use crate::rebased::sui::move_calls::utils;
+use crate::rebased::iota::move_calls::utils;
 
 use crate::rebased::utils::MoveType;
 use crate::rebased::Error;
