+++ identity_iota_core/src/rebased/iota/move_calls/identity/proposal.rs
@@ -1,7 +1,7 @@
 // Copyright 2020-2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::rebased::sui::move_calls::utils;
+use crate::rebased::iota::move_calls::utils;
 use crate::rebased::utils::MoveType;
 use crate::rebased::Error;
 use iota_sdk::rpc_types::OwnedObjectRef;
