+++ identity_iota_core/src/rebased/iota/move_calls/identity/send_asset.rs
@@ -12,7 +12,7 @@ use iota_sdk::types::TypeTag;
 use move_core_types::ident_str;
 
 use crate::rebased::proposals::SendAction;
-use crate::rebased::sui::move_calls;
+use crate::rebased::iota::move_calls;
 use crate::rebased::utils::MoveType;
 
 use self::move_calls::utils;
