+++ identity_iota_core/src/rebased/iota/move_calls/identity/upgrade.rs
@@ -9,37 +9,31 @@ use iota_sdk::types::transaction::ObjectArg;
 use iota_sdk::types::transaction::ProgrammableTransaction;
 use move_core_types::ident_str;
 
-use crate::rebased::sui::move_calls::utils;
+use crate::rebased::iota::move_calls::utils;
 
-pub(crate) fn propose_update(
+pub(crate) fn propose_upgrade(
   identity: OwnedObjectRef,
   capability: ObjectRef,
-  did_doc: impl AsRef<[u8]>,
   expiration: Option<u64>,
   package_id: ObjectID,
 ) -> Result<ProgrammableTransaction, anyhow::Error> {
   let mut ptb = ProgrammableTransactionBuilder::new();
   let cap_arg = ptb.obj(ObjectArg::ImmOrOwnedObject(capability))?;
-  let (delegation_token, borrow) = utils::get_controller_delegation(&mut ptb, cap_arg, package_id);
   let identity_arg = utils::owned_ref_to_shared_object_arg(identity, &mut ptb, true)?;
   let exp_arg = utils::option_to_move(expiration, &mut ptb, package_id)?;
-  let doc_arg = ptb.pure(did_doc.as_ref())?;
-  let clock = utils::get_clock_ref(&mut ptb);
 
   let _proposal_id = ptb.programmable_move_call(
     package_id,
     ident_str!("identity").into(),
-    ident_str!("propose_update").into(),
+    ident_str!("propose_upgrade").into(),
     vec![],
-    vec![identity_arg, delegation_token, doc_arg, exp_arg, clock],
+    vec![identity_arg, cap_arg, exp_arg],
   );
-
-  utils::put_back_delegation_token(&mut ptb, cap_arg, delegation_token, borrow, package_id);
 
   Ok(ptb.finish())
 }
 
-pub(crate) fn execute_update(
+pub(crate) fn execute_upgrade(
   identity: OwnedObjectRef,
   capability: ObjectRef,
   proposal_id: ObjectID,
@@ -47,20 +41,16 @@ pub(crate) fn execute_update(
 ) -> Result<ProgrammableTransaction, anyhow::Error> {
   let mut ptb = ProgrammableTransactionBuilder::new();
   let cap_arg = ptb.obj(ObjectArg::ImmOrOwnedObject(capability))?;
-  let (delegation_token, borrow) = utils::get_controller_delegation(&mut ptb, cap_arg, package_id);
   let proposal_id = ptb.pure(proposal_id)?;
   let identity_arg = utils::owned_ref_to_shared_object_arg(identity, &mut ptb, true)?;
-  let clock = utils::get_clock_ref(&mut ptb);
 
   let _ = ptb.programmable_move_call(
     package_id,
     ident_str!("identity").into(),
-    ident_str!("execute_update").into(),
+    ident_str!("execute_upgrade").into(),
     vec![],
-    vec![identity_arg, delegation_token, proposal_id, clock],
+    vec![identity_arg, cap_arg, proposal_id],
   );
-
-  utils::put_back_delegation_token(&mut ptb, cap_arg, delegation_token, borrow, package_id);
 
   Ok(ptb.finish())
 }
