+++ identity_iota_core/src/rebased/migration/alias.rs
@@ -19,7 +19,7 @@ use serde::Serialize;
 use crate::rebased::client::IdentityClient;
 use crate::rebased::client::IdentityClientReadOnly;
 use crate::rebased::client::IotaKeySignature;
-use crate::rebased::sui::move_calls;
+use crate::rebased::iota::move_calls;
 use crate::rebased::transaction::Transaction;
 use crate::rebased::transaction::TransactionOutput;
 use crate::rebased::utils::MoveType;
